Localized all strings and moved TableViewCell to a file.
Issue #90.
