+++ GiniVision/Classes/SupportedFormatsViewController.swift
@@ -10,27 +10,29 @@ import UIKit
 
 final class SupportedFormatsViewController: UITableViewController {
     
-    let reuseIdentifier = "reuseIdentifier"
+    typealias SupportedFormatCollectionSection = (title: String, items: [String], itemsImage:UIImage?, itemsImageBackgroundColor: UIColor)
+    
+    let supportedFormatsCellIdentifier = "SupportedFormatsCellIdentifier"
     let rowHeight: CGFloat = 70
     let sectionHeight: CGFloat = 70
-    var sections: [(title: String, items: [String], itemsImage:UIImage?, itemsImageBackgroundColor: UIColor)] = [
-        ("Folgende Formate werden unterstützt:",
-         ["Computer-erstellte Überweisungsträger und Rechnungen",
-          "Einseitige Bilder im jpg, png oder tif Format",
-          "PDF Dokumente von bis zu 10 Seiten"],
+    var sections: [SupportedFormatCollectionSection] = [
+        (NSLocalizedString("ginivision.supportedFormats.section.1.title", bundle: Bundle(for: GiniVision.self), comment: "title for supported formats section"),
+         [NSLocalizedString("ginivision.supportedFormats.section.1.item.1", bundle: Bundle(for: GiniVision.self), comment: "message for first item on supported formats section"),
+          NSLocalizedString("ginivision.supportedFormats.section.1.item.2", bundle: Bundle(for: GiniVision.self), comment: "message for second item on supported formats section"),
+          NSLocalizedString("ginivision.supportedFormats.section.1.item.3", bundle: Bundle(for: GiniVision.self), comment: "message for third item on supported formats section")],
          UIImage(named: "supportedFormatsIcon", in: Bundle(for: GiniVision.self), compatibleWith: nil),
          GiniConfiguration.sharedConfiguration.supportedFormatsIconColor),
-        ("Was nicht analysiert wird:",
-         ["Handschrift",
-          "Fotos von Bildschirme"],
+        (NSLocalizedString("ginivision.supportedFormats.section.2.title", bundle: Bundle(for: GiniVision.self), comment: "title for unsupported formats section"),
+         [NSLocalizedString("ginivision.supportedFormats.section.2.item.1", bundle: Bundle(for: GiniVision.self), comment: "message for first item on unsupported formats section"),
+          NSLocalizedString("ginivision.supportedFormats.section.2.item.2", bundle: Bundle(for: GiniVision.self), comment: "message for second item on unsupported formats section")],
          UIImage(named: "nonSupportedFormatsIcon", in: Bundle(for: GiniVision.self), compatibleWith: nil),
          GiniConfiguration.sharedConfiguration.nonSupportedFormatsIconColor)
     ]
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        title = "Unterstützte Formate"
-        tableView.register(SupportedFormatsTableViewCell.self, forCellReuseIdentifier: reuseIdentifier)
+        title = NSLocalizedString("ginivision.supportedFormats.title", bundle: Bundle(for: GiniVision.self), comment: "supported and unsupported formats screen title")
+        tableView.register(SupportedFormatsTableViewCell.self, forCellReuseIdentifier: supportedFormatsCellIdentifier)
         tableView.rowHeight = rowHeight
         tableView.tableFooterView = UIView()
         tableView.sectionHeaderHeight = sectionHeight
@@ -38,7 +40,7 @@ final class SupportedFormatsViewController: UITableViewController {
         tableView.backgroundColor = Colors.Gini.pearl
         tableView.alwaysBounceVertical = false
         
-        if #available(iOS 11.0, *) { // On iOS is .automatic by default.
+        if #available(iOS 11.0, *) { // On iOS is .automatic by default and it the transition to this view controller looks weird.
             tableView.contentInsetAdjustmentBehavior = .never
         }
     }
@@ -57,7 +59,7 @@ final class SupportedFormatsViewController: UITableViewController {
         let section = sections[indexPath.section]
         let item = section.items[indexPath.row]
         
-        let cell = tableView.dequeueReusableCell(withIdentifier: reuseIdentifier, for: indexPath) as! SupportedFormatsTableViewCell
+        let cell = tableView.dequeueReusableCell(withIdentifier: supportedFormatsCellIdentifier, for: indexPath) as! SupportedFormatsTableViewCell
         cell.textLabel?.text = item
         cell.imageView?.image = section.itemsImage
         cell.imageBackgroundView.backgroundColor = section.itemsImageBackgroundColor
@@ -75,42 +77,4 @@ final class SupportedFormatsViewController: UITableViewController {
         }
     }
 
-}
-
-final class SupportedFormatsTableViewCell: UITableViewCell {
-    
-    let imageViewSize = CGSize(width: 12, height: 12)
-    let imageBackgroundSize = CGSize(width: 22, height: 22)
-    
-    lazy var imageBackgroundView: UIView = {
-        let view = UIView(frame: CGRect(origin: .zero, size: self.imageBackgroundSize))
-        view.translatesAutoresizingMaskIntoConstraints = false
-        view.layer.cornerRadius = view.frame.width / 2
-        return view
-    }()
-    
-    override func draw(_ rect: CGRect) {
-        super.draw(rect)
-        
-        if let textLabel = textLabel {
-            textLabel.font = textLabel.font.withSize(14)
-            textLabel.numberOfLines = 0
-            textLabel.frame.origin = CGPoint(x: textLabel.frame.origin.x + imageBackgroundSize.width - imageViewSize.width, y: textLabel.frame.origin.y)
-        }
-    
-        if let imageView = imageView {
-            imageView.tintColor = .white
-            imageView.frame = CGRect(origin: CGPoint(x: imageView.frame.origin.x, y: (self.frame.height - imageViewSize.height) / 2), size: imageViewSize)
-            contentView.insertSubview(imageBackgroundView, belowSubview: imageView)
-            addConstraints()
-        }
-    }
-    
-    private func addConstraints() {
-        ConstraintUtils.addActiveConstraint(item: imageBackgroundView, attribute: .centerX, relatedBy: .equal, toItem: imageView!, attribute: .centerX, multiplier: 1.0, constant: 0)
-        ConstraintUtils.addActiveConstraint(item: imageBackgroundView, attribute: .centerY, relatedBy: .equal, toItem: imageView!, attribute: .centerY, multiplier: 1.0, constant: 0)
-
-        ConstraintUtils.addActiveConstraint(item: imageBackgroundView, attribute: .height, relatedBy: .equal, toItem: nil, attribute: .notAnAttribute, multiplier: 1.0, constant: 22)
-        ConstraintUtils.addActiveConstraint(item: imageBackgroundView, attribute: .width, relatedBy: .equal, toItem: nil, attribute: .notAnAttribute, multiplier: 1.0, constant: 22)
-    }
 }
